requirements.txt
psutil 
ping3
python-dotenv
